+++ shardtree/src/lib.rs
@@ -25,9 +25,10 @@
 use core::fmt::Debug;
 use either::Either;
 use incrementalmerkletree::frontier::Frontier;
+use incrementalmerkletree::Marking;
 use std::collections::{BTreeMap, BTreeSet};
 use std::sync::Arc;
-use tracing::trace;
+use tracing::{debug, trace};
 
 use incrementalmerkletree::{
     frontier::NonEmptyFrontier, Address, Hashable, Level, MerklePath, Position, Retention,
@@ -233,16 +234,17 @@ impl<
 
         let (append_result, position, checkpoint_id) =
             if let Some(subtree) = self.store.last_shard().map_err(ShardTreeError::Storage)? {
-                if subtree.root.is_complete() {
-                    let addr = subtree.root_addr;
-
-                    if addr.index() < Self::max_subtree_index() {
-                        LocatedTree::empty(addr.next_at_level()).append(value, retention)?
-                    } else {
-                        return Err(InsertionError::TreeFull.into());
+                match subtree.max_position() {
+                    // If the subtree is full, then construct a successor tree.
+                    Some(pos) if pos == subtree.root_addr.max_position() => {
+                        let addr = subtree.root_addr;
+                        if subtree.root_addr.index() < Self::max_subtree_index() {
+                            LocatedTree::empty(addr.next_at_level()).append(value, retention)?
+                        } else {
+                            return Err(InsertionError::TreeFull.into());
+                        }
                     }
-                } else {
-                    subtree.append(value, retention)?
+                    _ => subtree.append(value, retention)?,
                 }
             } else {
                 let root_addr = Address::from_parts(Self::subtree_level(), 0);
@@ -271,6 +273,7 @@ impl<
     ///
     /// This method may be used to add a checkpoint for the empty tree; note that
     /// [`Retention::Marked`] is invalid for the empty tree.
+    #[tracing::instrument(skip(self, frontier, leaf_retention))]
     pub fn insert_frontier(
         &mut self,
         frontier: Frontier<H, DEPTH>,
@@ -280,18 +283,19 @@ impl<
             self.insert_frontier_nodes(nonempty_frontier, leaf_retention)
         } else {
             match leaf_retention {
-                Retention::Ephemeral => Ok(()),
+                Retention::Ephemeral | Retention::Reference => Ok(()),
                 Retention::Checkpoint {
                     id,
-                    is_marked: false,
+                    marking: Marking::None | Marking::Reference,
                 } => self
                     .store
                     .add_checkpoint(id, Checkpoint::tree_empty())
                     .map_err(ShardTreeError::Storage),
-                Retention::Checkpoint {
-                    is_marked: true, ..
-                }
-                | Retention::Marked => Err(ShardTreeError::Insert(
+                Retention::Marked
+                | Retention::Checkpoint {
+                    marking: Marking::Marked,
+                    ..
+                } => Err(ShardTreeError::Insert(
                     InsertionError::MarkedRetentionInvalid,
                 )),
             }
@@ -301,6 +305,7 @@ impl<
     /// Add the leaf and ommers of the provided frontier as nodes within the subtree corresponding
     /// to the frontier's position, and update the cap to include the ommer nodes at levels greater
     /// than or equal to the shard height.
+    #[tracing::instrument(skip(self, frontier, leaf_retention))]
     pub fn insert_frontier_nodes(
         &mut self,
         frontier: NonEmptyFrontier<H>,
@@ -308,15 +313,24 @@ impl<
     ) -> Result<(), ShardTreeError<S::Error>> {
         let leaf_position = frontier.position();
         let subtree_root_addr = Address::above_position(Self::subtree_level(), leaf_position);
-        trace!("Subtree containing nodes: {:?}", subtree_root_addr);
 
-        let (updated_subtree, supertree) = self
+        let current_shard = self
             .store
             .get_shard(subtree_root_addr)
             .map_err(ShardTreeError::Storage)?
-            .unwrap_or_else(|| LocatedTree::empty(subtree_root_addr))
-            .insert_frontier_nodes(frontier, &leaf_retention)?;
+            .unwrap_or_else(|| LocatedTree::empty(subtree_root_addr));
 
+        trace!(
+            max_position = ?current_shard.max_position(),
+            subtree = ?current_shard,
+            "Current shard");
+        let (updated_subtree, supertree) =
+            current_shard.insert_frontier_nodes(frontier, &leaf_retention)?;
+        trace!(
+            max_position = ?updated_subtree.max_position(),
+            subtree = ?updated_subtree,
+            "Replacement shard"
+        );
         self.store
             .put_shard(updated_subtree)
             .map_err(ShardTreeError::Storage)?;
@@ -333,7 +347,7 @@ impl<
                 .map_err(ShardTreeError::Storage)?;
         }
 
-        if let Retention::Checkpoint { id, is_marked: _ } = leaf_retention {
+        if let Retention::Checkpoint { id, .. } = leaf_retention {
             trace!("Adding checkpoint {:?} at {:?}", id, leaf_position);
             self.store
                 .add_checkpoint(id, Checkpoint::at_position(leaf_position))
@@ -346,6 +360,7 @@ impl<
 
     /// Insert a tree by decomposing it into its `SHARD_HEIGHT` or smaller parts (if necessary)
     /// and inserting those at their appropriate locations.
+    #[tracing::instrument(skip(self, tree, checkpoints))]
     pub fn insert_tree(
         &mut self,
         tree: LocatedPrunableTree<H>,
@@ -359,6 +374,7 @@ impl<
             // for some inputs, and given that it is always correct to not insert an empty
             // subtree into `self`, we maintain the invariant by skipping empty subtrees.
             if subtree.root().is_empty() {
+                debug!("Subtree with root {:?} is empty.", subtree.root_addr);
                 continue;
             }
 
@@ -368,14 +384,15 @@ impl<
             let root_addr = Self::subtree_addr(subtree.root_addr.position_range_start());
 
             let contains_marked = subtree.root.contains_marked();
-            let (new_subtree, mut incomplete) = self
+            let current_shard = self
                 .store
                 .get_shard(root_addr)
                 .map_err(ShardTreeError::Storage)?
-                .unwrap_or_else(|| LocatedTree::empty(root_addr))
-                .insert_subtree(subtree, contains_marked)?;
+                .unwrap_or_else(|| LocatedTree::empty(root_addr));
+            let (replacement_shard, mut incomplete) =
+                current_shard.insert_subtree(subtree, contains_marked)?;
             self.store
-                .put_shard(new_subtree)
+                .put_shard(replacement_shard)
                 .map_err(ShardTreeError::Storage)?;
             all_incomplete.append(&mut incomplete);
         }
@@ -396,8 +413,8 @@ impl<
             root_addr: Address,
             root: &PrunableTree<H>,
         ) -> Option<(PrunableTree<H>, Position)> {
-            match root {
-                Tree(Node::Parent { ann, left, right }) => {
+            match &root.0 {
+                Node::Parent { ann, left, right } => {
                     let (l_addr, r_addr) = root_addr.children().unwrap();
                     go(r_addr, right).map_or_else(
                         || {
@@ -426,13 +443,13 @@ impl<
                         },
                     )
                 }
-                Tree(Node::Leaf { value: (h, r) }) => Some((
+                Node::Leaf { value: (h, r) } => Some((
                     Tree(Node::Leaf {
                         value: (h.clone(), *r | RetentionFlags::CHECKPOINT),
                     }),
                     root_addr.max_position(),
                 )),
-                Tree(Node::Nil) => None,
+                Node::Nil | Node::Pruned => None,
             }
         }
 
@@ -478,6 +495,7 @@ impl<
         Ok(true)
     }
 
+    #[tracing::instrument(skip(self))]
     fn prune_excess_checkpoints(&mut self) -> Result<(), ShardTreeError<S::Error>> {
         let checkpoint_count = self
             .store
@@ -559,12 +577,19 @@ impl<
 
             // Prune each affected subtree
             for (subtree_addr, positions) in clear_positions.into_iter() {
-                let cleared = self
+                let to_clear = self
                     .store
                     .get_shard(subtree_addr)
-                    .map_err(ShardTreeError::Storage)?
-                    .map(|subtree| subtree.clear_flags(positions));
-                if let Some(cleared) = cleared {
+                    .map_err(ShardTreeError::Storage)?;
+
+                if let Some(to_clear) = to_clear {
+                    let pre_clearing_max_position = to_clear.max_position();
+                    let cleared = to_clear.clear_flags(positions);
+
+                    // Clearing flags should not modify the max position of leaves represented
+                    // in the shard.
+                    assert!(cleared.max_position() == pre_clearing_max_position);
+
                     self.store
                         .put_shard(cleared)
                         .map_err(ShardTreeError::Storage)?;
@@ -740,8 +765,8 @@ impl<
         // roots.
         truncate_at: Position,
     ) -> Result<(H, Option<PrunableTree<H>>), ShardTreeError<S::Error>> {
-        match &cap.root {
-            Tree(Node::Parent { ann, left, right }) => {
+        match &cap.root.0 {
+            Node::Parent { ann, left, right } => {
                 match ann {
                     Some(cached_root) if target_addr.contains(&cap.root_addr) => {
                         Ok((cached_root.as_ref().clone(), None))
@@ -815,7 +840,7 @@ impl<
                     }
                 }
             }
-            Tree(Node::Leaf { value }) => {
+            Node::Leaf { value } => {
                 if truncate_at >= cap.root_addr.position_range_end()
                     && target_addr.contains(&cap.root_addr)
                 {
@@ -840,7 +865,7 @@ impl<
                     ))
                 }
             }
-            Tree(Node::Nil) => {
+            Node::Nil | Node::Pruned => {
                 if cap.root_addr == target_addr
                     || cap.root_addr.level() == ShardTree::<S, DEPTH, SHARD_HEIGHT>::subtree_level()
                 {
@@ -850,7 +875,7 @@ impl<
                     Ok((
                         root.clone(),
                         if truncate_at >= cap.root_addr.position_range_end() {
-                            // return the compute root as a new leaf to be cached if it contains no
+                            // return the computed root as a new leaf to be cached if it contains no
                             // empty hashes due to truncation
                             Some(Tree::leaf((root, RetentionFlags::EPHEMERAL)))
                         } else {
@@ -1284,21 +1309,24 @@ impl<
 
 #[cfg(test)]
 mod tests {
+    use std::convert::Infallible;
+
     use assert_matches::assert_matches;
     use proptest::prelude::*;
 
     use incrementalmerkletree::{
-        frontier::NonEmptyFrontier,
+        frontier::{Frontier, NonEmptyFrontier},
         testing::{
             arb_operation, check_append, check_checkpoint_rewind, check_operations,
             check_remove_mark, check_rewind_remove_mark, check_root_hashes,
             check_witness_consistency, check_witnesses, complete_tree::CompleteTree, CombinedTree,
             SipHashable,
         },
-        Address, Hashable, Level, Position, Retention,
+        Address, Hashable, Level, Marking, MerklePath, Position, Retention,
     };
 
     use crate::{
+        error::{QueryError, ShardTreeError},
         store::memory::MemoryShardStore,
         testing::{
             arb_char_str, arb_shardtree, check_shard_sizes, check_shardtree_insertion,
@@ -1396,11 +1424,100 @@ mod tests {
                 'd'.to_string(),
                 Retention::Checkpoint {
                     id: 11,
-                    is_marked: false
+                    marking: Marking::None
                 },
             ),
             Ok(()),
         );
+    }
+
+    #[test]
+    fn avoid_pruning_reference() {
+        fn test_with_marking(
+            frontier_marking: Marking,
+        ) -> Result<MerklePath<String, 6>, ShardTreeError<Infallible>> {
+            let mut tree = ShardTree::<MemoryShardStore<String, usize>, 6, 3>::new(
+                MemoryShardStore::empty(),
+                5,
+            );
+
+            let frontier_end = Position::from((1 << 3) - 3);
+            let mut f0 = Frontier::<String, 6>::empty();
+            for c in 'a'..='f' {
+                f0.append(c.to_string());
+            }
+
+            let frontier = Frontier::from_parts(
+                frontier_end,
+                "f".to_owned(),
+                vec!["e".to_owned(), "abcd".to_owned()],
+            )
+            .unwrap();
+
+            // Insert a frontier two leaves from the end of the first shard, checkpointed,
+            // with the specified marking.
+            tree.insert_frontier(
+                frontier,
+                Retention::Checkpoint {
+                    id: 1,
+                    marking: frontier_marking,
+                },
+            )?;
+
+            // Insert a few leaves beginning at the subsequent position, so as to cross the shard
+            // boundary.
+            tree.batch_insert(
+                frontier_end + 1,
+                ('g'..='j')
+                    .into_iter()
+                    .map(|c| (c.to_string(), Retention::Ephemeral)),
+            )?;
+
+            // Trigger pruning by adding 5 more checkpoints
+            for i in 2..7 {
+                tree.checkpoint(i).unwrap();
+            }
+
+            // Insert nodes that require the pruned nodes for witnessing
+            tree.batch_insert(
+                frontier_end - 1,
+                ('e'..='f')
+                    .into_iter()
+                    .map(|c| (c.to_string(), Retention::Marked)),
+            )?;
+
+            // Compute the witness
+            tree.witness_at_checkpoint_id(frontier_end, &6)
+        }
+
+        // If we insert the frontier with Marking::None, the frontier nodes are treated
+        // as ephemeral nodes and are pruned, leaving an incomplete tree.
+        assert_matches!(
+            test_with_marking(Marking::None),
+            Err(ShardTreeError::Query(QueryError::TreeIncomplete(_)))
+        );
+
+        // If we insert the frontier with Marking::Reference, the frontier nodes will
+        // not be pruned on completion of the subtree, and thus we'll be able to compute
+        // the witness.
+        let expected_witness = MerklePath::from_parts(
+            [
+                "e",
+                "gh",
+                "abcd",
+                "ij______",
+                "________________",
+                "________________________________",
+            ]
+            .iter()
+            .map(|s| s.to_string())
+            .collect(),
+            Position::from(5),
+        )
+        .unwrap();
+
+        let witness = test_with_marking(Marking::Reference).unwrap();
+        assert_eq!(witness, expected_witness);
     }
 
     // Combined tree tests